Add modules and exports
