+++ index.css
@@ -1,0 +1,97 @@
+body {
+  margin: 0;
+  font-family: "Times New Roman", Times, serif;
+}
+
+.top-header {
+  margin: 15px;
+  padding: 0;
+}
+
+.navbar {
+  margin: 0;
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  padding: 0 35px;
+  border: 5px solid rgb(0, 0, 0);
+  background-color: #ddd;
+}
+
+.nav-name {
+  padding-right: 400px;
+}
+
+a {
+  text-decoration: none;
+  color: rgb(0, 0, 0);
+}
+
+.date-container {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin-right: 25px;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+table {
+  border: 4px solid rgb(0, 0, 0);
+}
+
+.delete {
+  margin-left: 150px;
+  background-color: #fff;
+}
+
+hr {
+  width: 100px;
+  margin: 35px 0;
+}
+
+#form {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.form-group {
+  margin: 10px 0;
+}
+
+.submit {
+  align-self: flex-end;
+}
+
+#book-list {
+  display: flex;
+  justify-content: center;
+  align-self: center;
+}
+
+#add-new {
+  display: flex;
+  justify-content: center;
+  align-self: center;
+}
+
+#contact {
+  display: flex;
+  justify-content: center;
+  align-self: center;
+}
+
+footer {
+  position: absolute;
+  bottom: 0;
+  width: 96%;
+  margin: 25px;
+  border: 5px solid rgb(0, 0, 0);
+}
