+++ index.js
@@ -1,0 +1,10 @@
+/* eslint-disable no-unused-vars */
+import { Book } from "./Modules/Class-book";
+import { showBooks } from "./Modules/DOM-manip";
+import { Store } from "./Modules/Store";
+import { bookAdd } from "./Modules/ShowBookEvents";
+import { bookDelete } from "./Modules/ShowBookEvents";
+import { bookShow } from "./Modules/ShowBookEvents";
+import { showAdd } from "./Modules/Events";
+import { showContactPage } from "./Modules/Events";
+import { showList } from "./Modules/Events";
